+++ app/serializers/dc20/summon_serializer.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Dc20
+  class SummonSerializer < ApplicationSerializer
+    ATTRIBUTES = %i[id name data].freeze
+
+    attributes(*ATTRIBUTES)
+  end
+end