+++ config/initializers/container.rb
@@ -68,6 +68,8 @@ module Charkeeper
     register('commands.characters_context.dc20.wild_forms.create') { CharactersContext::Dc20::WildForms::CreateCommand.new }
     register('commands.characters_context.dc20.wild_forms.update') { CharactersContext::Dc20::WildForms::UpdateCommand.new }
 
+    register('commands.characters_context.dc20.summons.create') { CharactersContext::Dc20::Summons::CreateCommand.new }
+
     register('commands.characters_context.dc20.create') { CharactersContext::Dc20::CreateCommand.new }
     register('commands.characters_context.dc20.update') { CharactersContext::Dc20::UpdateCommand.new }