+++ config/routes.rb
@@ -209,6 +209,7 @@ Rails.application.routes.draw do
         scope module: :characters do
           resources :ancestries, only: %i[index]
           resources :wild_forms, only: %i[index create update destroy]
+          resources :summons, only: %i[index create update destroy]
           resources :spells, only: %i[index create update destroy] do
             post :clear, on: :collection
           end
@@ -226,6 +227,7 @@ Rails.application.routes.draw do
       resources :ancestries, only: %i[index]
       resources :spells, only: %i[index]
       resources :wild_forms, only: %i[index]
+      resources :summons, only: %i[index]
     end
 
     namespace :nimble do