+++ db/seeds_prod.rb
@@ -53,7 +53,7 @@ end
   'talents/barbarian.json', 'talents/bard.json', 'talents/champion.json', 'talents/cleric.json', 'talents/commander.json',
   'talents/druid.json', 'talents/hunter.json', 'talents/monk.json', 'talents/rogue.json', 'talents/sorcerer.json',
   'talents/champion.json',
-  'talents/spellblade.json', 'talents/warlock.json', 'talents/wizard.json',
+  'talents/spellblade.json', 'talents/warlock.json', 'talents/wizard.json', 'summons.json',
   'ancestry.json', 'ancestry_2.json', 'ancestry_3.json', 'spells.json', 'maneuvers.json', 'talents.json', 'wild_forms.json'
 ].each do |filename|
   response = Net::HTTP.get(URI("https://raw.githubusercontent.com/kortirso/charkeeper_data/refs/heads/master/dc20/#{filename}"))